refactor(commands): extract classification tests to sibling file
